Organized db pool aquiring
